opt: 更新统计信息格式
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.8@32626 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/utilCache.cpp
@@ -3168,7 +3168,7 @@ namespace engine
                    OSS_NEWLINE
                    "Unit Name         : %s"OSS_NEWLINE
                    "Total Time(Sec)   : %u"OSS_NEWLINE
-                   "Cache Total Sz(MB): %u"OSS_NEWLINE
+                   "Cache Total Sz(MB): %.2f"OSS_NEWLINE
                    "Cache Free Sz(MB) : %.2f"OSS_NEWLINE
                    "Total Page        : %u"OSS_NEWLINE
                    "Dirty Page        : %u"OSS_NEWLINE
@@ -3191,7 +3191,7 @@ namespace engine
                    "Recycle Speed     : %.2f /s"OSS_NEWLINE,
                    _pCacheFile->getFileName(),
                    diff / 1000,
-                   _pMgr->totalSize() / 1048576,
+                   (FLOAT64)_pMgr->totalSize() / 1048576,
                    (FLOAT64)_pMgr->freeSize() / 1048576,
                    statTotalPage,
                    statDirtyPage,
